Class Usuario e Class Transacao criadas
