--- src/main/java/org/example/LoginScreen.java
@@ -1,230 +1,0 @@
-package org.example;
-
-import javax.swing.*;
-import java.awt.*;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-
-public class LoginScreen extends JFrame {
-    private JTextField userField;
-    private JPasswordField passwordField;
-
-    public LoginScreen() {
-        setTitle("Login");
-        setSize(900, 600);
-        setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-        setLocationRelativeTo(null);
-
-        JPanel mainPanel = new JPanel();
-        mainPanel.setLayout(new BorderLayout());
-        mainPanel.setBackground(new Color(240, 240, 240));
-
-        // Painel para o formulário de login
-        JPanel formPanel = new JPanel();
-        formPanel.setLayout(new GridBagLayout());
-        formPanel.setBackground(new Color(240, 240, 240));
-        GridBagConstraints gbc = new GridBagConstraints();
-        gbc.insets = new Insets(10, 10, 10, 10);
-
-        // Adicionando um título
-        JLabel titleLabel = new JLabel("Gerenciado Financeiro");
-        titleLabel.setFont(new Font("Arial", Font.BOLD, 24));
-        titleLabel.setForeground(new Color(50, 50, 50));
-        gbc.gridx = 0;
-        gbc.gridy = 0;
-        gbc.gridwidth = 2;
-        gbc.anchor = GridBagConstraints.CENTER;
-        formPanel.add(titleLabel, gbc);
-
-        // Campo de Email
-        gbc.gridx = 0;
-        gbc.gridy = 1;
-        gbc.gridwidth = 1;
-        gbc.anchor = GridBagConstraints.LINE_END;
-        formPanel.add(new JLabel("Email:"), gbc);
-
-        gbc.gridx = 1;
-        gbc.gridy = 1;
-        gbc.anchor = GridBagConstraints.LINE_START;
-        userField = new JTextField(20);
-        formPanel.add(userField, gbc);
-
-        // Campo de senha
-        gbc.gridx = 0;
-        gbc.gridy = 2;
-        gbc.anchor = GridBagConstraints.LINE_END;
-        formPanel.add(new JLabel("Senha:"), gbc);
-
-        gbc.gridx = 1;
-        gbc.gridy = 2;
-        gbc.anchor = GridBagConstraints.LINE_START;
-        passwordField = new JPasswordField(20);
-        formPanel.add(passwordField, gbc);
-
-        // Botão de login
-        JButton loginButton = new JButton("ENTRAR");
-        loginButton.setFont(new Font("Arial", Font.BOLD, 14));
-        loginButton.setBackground(new Color(50, 150, 250));
-        loginButton.setForeground(Color.WHITE);
-        loginButton.setFocusPainted(false);
-        loginButton.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                String user = userField.getText();
-                String password = new String(passwordField.getPassword());
-
-                if (user.isEmpty() || password.isEmpty()) {
-                    JOptionPane.showMessageDialog(LoginScreen.this, "Por favor, preencha todos os campos.", "Erro", JOptionPane.ERROR_MESSAGE);
-                } else {
-                    //logica
-                    new FinanceManagerScreen(user).setVisible(true);
-                    dispose();
-                }
-            }
-        });
-
-
-        // Botão de cadastrar
-        JButton registerButton = new JButton("CADASTRAR");
-        registerButton.setFont(new Font("Arial", Font.BOLD, 14));
-        registerButton.setBackground(new Color(50, 150, 250));
-        registerButton.setForeground(Color.WHITE);
-        registerButton.setFocusPainted(false);
-        registerButton.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                new RegisterScreen();
-                dispose();
-            }
-        });
-
-        // Painel para os botões
-        JPanel buttonPanel = new JPanel();
-        buttonPanel.setBackground(new Color(240, 240, 240));
-        buttonPanel.add(loginButton);
-        buttonPanel.add(registerButton);
-
-        gbc.gridx = 0;
-        gbc.gridy = 3;
-        gbc.gridwidth = 2;
-        gbc.anchor = GridBagConstraints.CENTER;
-        formPanel.add(buttonPanel, gbc);
-
-        mainPanel.add(formPanel, BorderLayout.CENTER);
-        add(mainPanel);
-        setVisible(true);
-    }
-}
-
-class RegisterScreen extends JFrame {
-    private JTextField nameField;
-    private JTextField userField;
-    private JPasswordField passwordField;
-    private JTextField emailField;
-
-    public RegisterScreen() {
-        setTitle("Cadastro");
-        setSize(900, 600);
-        setDefaultCloseOperation(JFrame.DISPOSE_ON_CLOSE);
-        setLocationRelativeTo(null);
-
-        JPanel mainPanel = new JPanel();
-        mainPanel.setLayout(new BorderLayout());
-        mainPanel.setBackground(new Color(240, 240, 240));
-
-        // Painel para o formulário de cadastro
-        JPanel formPanel = new JPanel();
-        formPanel.setLayout(new GridBagLayout());
-        formPanel.setBackground(new Color(240, 240, 240));
-        GridBagConstraints gbc = new GridBagConstraints();
-        gbc.insets = new Insets(10, 10, 10, 10);
-
-        // Adicionando um título
-        JLabel titleLabel = new JLabel("Cadastro de Usuário");
-        titleLabel.setFont(new Font("Arial", Font.BOLD, 24));
-        titleLabel.setForeground(new Color(50, 50, 50));
-        gbc.gridx = 0;
-        gbc.gridy = 0;
-        gbc.gridwidth = 2;
-        gbc.anchor = GridBagConstraints.CENTER;
-        formPanel.add(titleLabel, gbc);
-
-        // Campo de nome
-        gbc.gridx = 0;
-        gbc.gridy = 1;
-        gbc.gridwidth = 1;
-        gbc.anchor = GridBagConstraints.LINE_END;
-        formPanel.add(new JLabel("Nome:"), gbc);
-
-        gbc.gridx = 1;
-        gbc.gridy = 1;
-        gbc.anchor = GridBagConstraints.LINE_START;
-        nameField = new JTextField(20);
-        formPanel.add(nameField, gbc);
-
-        // Campo de e-mail
-        gbc.gridx = 0;
-        gbc.gridy = 2;
-        gbc.anchor = GridBagConstraints.LINE_END;
-        formPanel.add(new JLabel("E-mail:"), gbc);
-
-        gbc.gridx = 1;
-        gbc.gridy = 2;
-        gbc.anchor = GridBagConstraints.LINE_START;
-        emailField = new JTextField(20);
-        formPanel.add(emailField, gbc);
-
-        // Campo de senha
-        gbc.gridx = 0;
-        gbc.gridy = 3;
-        gbc.anchor = GridBagConstraints.LINE_END;
-        formPanel.add(new JLabel("Senha:"), gbc);
-
-        gbc.gridx = 1;
-        gbc.gridy = 3;
-        gbc.anchor = GridBagConstraints.LINE_START;
-        passwordField = new JPasswordField(20);
-        formPanel.add(passwordField, gbc);
-
-        // Botão de cadastrar
-        JButton registerButton = new JButton("CADASTRAR");
-        registerButton.setFont(new Font("Arial", Font.BOLD, 14));
-        registerButton.setBackground(new Color(50, 150, 250));
-        registerButton.setForeground(Color.WHITE);
-        registerButton.setFocusPainted(false);
-        registerButton.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                String name = nameField.getText();
-                String email = emailField.getText();
-                String password = new String(passwordField.getPassword());
-
-
-                if (name.isEmpty() || password.isEmpty() || email.isEmpty()) {
-                    JOptionPane.showMessageDialog(RegisterScreen.this, "Por favor, preencha todos os campos.", "Erro", JOptionPane.ERROR_MESSAGE);
-                } else {
-                    // lógica
-                    JOptionPane.showMessageDialog(RegisterScreen.this, "Cadastro realizado com sucesso!", "Sucesso", JOptionPane.INFORMATION_MESSAGE);
-                    new LoginScreen();
-                    dispose();
-                }
-            }
-        });
-
-        // Painel para os botões
-        JPanel buttonPanel = new JPanel();
-        buttonPanel.setBackground(new Color(240, 240, 240));
-        buttonPanel.add(registerButton);
-
-        gbc.gridx = 0;
-        gbc.gridy = 5;
-        gbc.gridwidth = 2;
-        gbc.anchor = GridBagConstraints.CENTER;
-        formPanel.add(buttonPanel, gbc);
-
-        mainPanel.add(formPanel, BorderLayout.CENTER);
-        add(mainPanel);
-        setVisible(true);
-
-    }
-}
