+++ src/org/example/Main.java
@@ -9,7 +9,6 @@ public class Main {
             public void run() {
                 new LoginScreen();
             }
-
         });
     }
 }
